Rename human format to friendly and make it default.
